Criação da autenticação, sessões e middleware

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\LoginController;
 use App\Http\Controllers\PetsController;
 use App\Http\Controllers\PetsAdminController;
 use App\Http\Controllers\SolicitanteController;
@@ -21,7 +22,6 @@ Route::get('/', function () {
 })->name('home');
 
 Route::get('/queroadotar', [PetsController::class, 'RetornaPets'])->name('queroadotar');
-Route::get('/queroadotar/filtrar', [PetsController::class, 'filtrarPet'])->name('filtrarpet');
 
 Route::get('/adotar', [PetsController::class, 'mostrar'])->name('adotar');
 
@@ -33,22 +33,23 @@ Route::prefix('formulario')->group(function () {
 
 
 Route::prefix('cadastrar')->group(function () {
-    Route::get('/', [PetsAdminController::class, 'index'])->name('cadastrar');
-    Route::get('/create', [PetsAdminController::class, 'create'])->name('cadastrar-create');
-    Route::post('/', [PetsAdminController::class, 'store'])->name('cadastrar-store');
+    Route::get('/', [PetsAdminController::class, 'index'])->name('cadastrar')->middleware('auth');
+    Route::get('/create', [PetsAdminController::class, 'create'])->name('cadastrar-create')->middleware('auth');
+    Route::post('/', [PetsAdminController::class, 'store'])->name('cadastrar-store')->middleware('auth');
 });
 
 Route::prefix('painel')->group(function () {
-    Route::get('/', [PetsAdminController::class, 'painelPets'])->name('painel');
-    Route::get('/{id}/editar', [PetsAdminController::class, 'edit'])->where('id', '[0-9]+')->name('painel-editar');
-    Route::put('/{id}', [PetsAdminController::class, 'update'])->where('id', '[0-9]+')->name('painel-update');
-    Route::delete('/{id}', [PetsAdminController::class, 'destroy'])->where('id', '[0-9]+')->name('painel-destroy');
+    Route::get('/', [PetsAdminController::class, 'painelPets'])->name('painel')->middleware('auth');
+    Route::get('/solicitantes', [PetsAdminController::class, 'painelSolicitacoes'])->name('painel-solicitacoes')->middleware('auth');
+    Route::get('/filtrar', [PetsAdminController::class, 'filtrar'])->name('painel-filtrar')->middleware('auth');
+    Route::get('/{id}/editar', [PetsAdminController::class, 'edit'])->where('id', '[0-9]+')->name('painel-editar')->middleware('auth');
+    Route::put('/{id}', [PetsAdminController::class, 'update'])->where('id', '[0-9]+')->name('painel-update')->middleware('auth');
+    Route::delete('/{id}', [PetsAdminController::class, 'destroy'])->where('id', '[0-9]+')->name('painel-destroy')->middleware('auth');
 });
 
-
-Route::get('/login', function () {
-    return view('login');
-})->name('login');
+Route::view('/login', 'login')->name('login.form');
+Route::post('/auth', [LoginController::class, 'auth'])->name('login.auth');
+Route::get('/logout', [LoginController::class,'logout'])->name('login.logout');
 
 Route::get('/recuperarsenha', function () {
     return view('recuperarsenha');
